refactor(qvism): reduce type checking time
* StyleObject return type in animation utility; spreading complex object is expensive
* Refactor StyleObject type definition for better type inference
* Provide define utilities to ensure type validity
