+++ packages/css/seed.css
@@ -582,7 +582,7 @@
   overflow: hidden;
 }
 
-.seed-avatar__image:not([data-loading-status="loaded"]) {
+.seed-avatar__image:not([data-loading-state="loaded"]) {
   display: none;
 }
 
@@ -597,7 +597,7 @@
   overflow: hidden;
 }
 
-.seed-avatar__fallback[data-loading-status="loaded"] {
+.seed-avatar__fallback[data-loading-state="loaded"] {
   display: none;
 }
