+++ packages/qvism-preset/src/recipes/action-sheet-item.ts
@@ -1,5 +1,5 @@
 import { actionSheetItem as vars, actionSheet as rootVars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { active, pseudo } from "../utils/pseudo";
 
 const actionSheetItem = defineRecipe({
