+++ packages/qvism-preset/src/recipes/action-sheet.ts
@@ -3,7 +3,7 @@ import {
   actionSheet as vars,
 } from "@seed-design/css/vars/component";
 import { enterAnimation, exitAnimation } from "../utils/animation";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { active, not, open, pseudo } from "../utils/pseudo";
 
 const actionSheet = defineRecipe({
