+++ packages/qvism-preset/src/recipes/badge.ts
@@ -1,5 +1,5 @@
 import { badge as vars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 
 const badge = defineRecipe({
   name: "badge",
