+++ packages/qvism-preset/src/recipes/fab.ts
@@ -1,6 +1,6 @@
 import { fab as vars } from "@seed-design/css/vars/component";
 
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { active, disabled, focus, pseudo } from "../utils/pseudo";
 import { onlyIcon } from "../utils/icon";
