+++ packages/qvism-preset/src/recipes/identity-placeholder.ts
@@ -1,5 +1,5 @@
 import { identityPlaceholder as vars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 
 const identityPlaceholder = defineRecipe({
   name: "identity-placeholder",
