+++ packages/qvism-preset/src/recipes/inline-banner.ts
@@ -1,5 +1,5 @@
 import { inlineBanner as vars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { pseudo } from "../utils/pseudo";
 import { prefixIcon, suffixIcon } from "../utils/icon";
