+++ packages/qvism-preset/src/recipes/progress-circle.ts
@@ -1,5 +1,5 @@
 import { progressCircle as vars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 
 const progressCircle = defineRecipe({
   name: "progress-circle",
