+++ packages/qvism-preset/src/recipes/segmented-control.ts
@@ -2,7 +2,7 @@ import {
   segmentedControlItem as itemVars,
   segmentedControl as vars,
 } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { active, checked, disabled, not, pseudo } from "../utils/pseudo";
 
 const segmentedControl = defineRecipe({
