+++ packages/qvism-preset/src/recipes/select-box.ts
@@ -1,5 +1,5 @@
 import { selectBox as vars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { active, checked, pseudo } from "../utils/pseudo";
 
 export const selectBoxGroup = defineRecipe({
