+++ packages/qvism-preset/src/recipes/skeleton.ts
@@ -1,5 +1,5 @@
 import { skeleton as vars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 
 const skeleton = defineRecipe({
   name: "skeleton",
