+++ packages/qvism-preset/src/recipes/tabs.ts
@@ -1,5 +1,5 @@
 import { tablist as vars, tab as triggerVars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { disabled, not, pseudo, selected } from "../utils/pseudo";
 
 const tabs = defineRecipe({
