+++ packages/qvism-preset/src/recipes/text-field.ts
@@ -1,5 +1,5 @@
 import { textField as vars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { pseudo, focus, disabled, not, readOnly, invalid } from "../utils/pseudo";
 
 const textField = defineRecipe({
