+++ packages/qvism-preset/src/stackflow/app-bar.ts
@@ -1,5 +1,5 @@
 import { topNavigation as vars } from "@seed-design/css/vars/component";
-import { defineRecipe } from "../utils/define-recipe";
+import { defineRecipe } from "../utils/define";
 import { fadeFromBottomAndroidAnimations, iOSAnimations } from "./animation";
 import {
   idle,
