--- packages/qvism-preset/src/utils/define-keyframes.ts
@@ -1,7 +1,0 @@
-import type { KeyframeDefinition } from "@seed-design/qvism-core";
-
-export function defineKeyframes(
-  definition: Record<string, KeyframeDefinition>,
-): Record<string, KeyframeDefinition> {
-  return definition;
-}
