--- packages/qvism-preset/src/utils/define-recipe.ts
@@ -1,5 +1,0 @@
-import type { SlotRecipeDefinition } from "@seed-design/qvism-core";
-
-export function defineRecipe<T extends SlotRecipeDefinition<string, any>>(definition: T): T {
-  return definition;
-}
